LPD-33938 Move parent folder
