+++ modules/test/playwright/tests/workspaces/liferay-workspace-marketplace/helpers/MarketplaceHelper.ts
@@ -77,6 +77,7 @@ export class MarketplaceHelper {
 		}
 		catch (error) {
 			console.error('Error when trying to create account', error);
+
 			throw error;
 		}
 	}
